Add steps to configure pinry before building.
I followed these instructions and was able to get pinry running but couldn't create a new user from the web and tear down the container and read the pinry docs for the initial configuration. There may be a better way to call createsuperuser from outside the container after start that would work better for this guide.